done expand file env and separeate file
