+++ Makefile
@@ -12,13 +12,15 @@ SRCS_NAME = main.c \
 			envlist.c \
 			execdata.c \
 			execdata_utils.c \
-			execution_setdata.c \
+			setdata_cmdline_redirection.c \
+			setdata_heredoc_cmdtype.c \
 			execution_start.c \
 			execution_utils.c \
 			execution_utils_libft.c \
 			execution_utils_list.c \
 			expansion_cmdlist.c \
 			expansion_iolist.c \
+			expansion_io_utils.c \
 			expansion_utils1.c \
 			expansion_utils2.c \
 			minishell_loop.c \
@@ -29,7 +31,7 @@ SRCS_NAME = main.c \
 			tokenize_utils.c \
 			wrapper1.c \
 			wrapper2.c \
-			wrapper3.c			
+			wrapper3.c 
 
 SRCS = ${addprefix ${SRCS_DIR}, ${SRCS_NAME}}
