+++ includes/execute.h
@@ -17,8 +17,10 @@ void		execute_command(t_execdata *data);
 int			execute_loop(t_execdata *data);
 void		execute_start(t_execdata *data);
 
-//execution_setdata.c
+//setdata_cmdline_redirection.c
 int			setdata_cmdline_redirect(t_execdata *data);
+
+//setdata_heredoc_cmdtype.c
 void		setdata_heredoc_cmdtype(t_execdata *data);
 
 //command_*.c
@@ -67,9 +69,11 @@ pid_t		xfork(void);
 //minishell_loop.c
 void		minishell_loop(char **envp);
 
-//after
-void		clear_quot_filename(char **filename, char **filequot);
+//expansion_iolist.c
 int			serch_env_iolist(t_iolist *iolist, t_envlist *envlist);
+
+//expansion_io_utils.c
+void		clear_quot_filename(char **filename, char **filequot);
 void		expansion_key_io(char **line, \
 				t_envlist *envlist, char *doll_ptr);
 char		*ft_strdoll(const char *s);
